Improve printing and docs for evalPQ.

+++ panoptic_parts/evaluation/eval_PartPQ.py
@@ -50,7 +50,7 @@ def pred_reader_fn(fp_pred):
   return pan_classes, pan_inst_ids, parts_output
 
 
-def evaluate(eval_spec_path, basepath_gt, basepath_pred, save_dir):
+def evaluate(eval_spec_path, basepath_gt, basepath_pred, save_dir=None):
   spec = PPQEvalSpec(eval_spec_path)
 
   dst_name = spec._dspec.dataset_name
@@ -58,20 +58,26 @@ def evaluate(eval_spec_path, basepath_gt, basepath_pred, save_dir):
 
   results = evaluate_PartPQ_multicore(spec, filepaths_pairs, pred_reader_fn)
 
+  print()
   print(*map(lambda d: ', '.join(map(lambda t: f'{t[0]}: {t[1]:.3f}', d.items())),
              results[0]),
         sep='\n')
   print(*map(lambda t: f'{t[0]:15} ' + ', '.join(map(lambda t: f'{t[0]}: {t[1]:.3f}', t[1].items())),
              zip(spec.eval_sid2scene_label.values(), results[1].values())),
         sep='\n')
+  print()
 
   if save_dir is not None:
     if not op.exists(save_dir):
       print("Creating output directory at {}".format(save_dir))
       os.mkdir(save_dir)
-    with open(op.join(save_dir, 'partpq_results.json'), 'w') as fp:
+    filepath_output = op.join(save_dir, f'{op.basename(eval_spec_path)[:-14]}_results.json')
+    if op.exists(filepath_output):
+      if input(f'Overwrite existing file {filepath_output} (y/n): ') == 'n':
+        exit()
+    with open(filepath_output, 'w') as fp:
       json.dump(results, fp)
-    print("Results saved in {}".format(op.join(save_dir, 'partpq_results.json')))
+    print("Results saved in {}".format(filepath_output))
 
 
 if __name__ == '__main__':
